Add gn_isolate_map.pyl to trybot_analyze_config.json.
Previously, if you uploaded a CL that only changed
//testing/buildbot/gn_isolate_map.pyl, then no tests would
be run: analyze would not think anything had changed. This
should really be fixed by marking it as a data dependency
of all of the test targets, but that's hard to do without
the generic build wrapper framework.

Besides, I'm trying to make the file be obsolete.

In the meantime, this'll mark it so that we do trigger tests
and make sure changes to the file don't break things.

Bug: 816629
Change-Id: If9f65e743d276582720860893dcb877e75ba1844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354779
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Erik Staab <estaab@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#798003}
GitOrigin-RevId: 2c022722b009d58c3bf683855be2a4711c4f8830