Mark cronet_unittests_ios as a 'console_test_launcher'.
It looks like this binary was originally declared as a 'raw'
test (meaning that we didn't know how to do anything other than
look at the exit code to determine pass/fail), but that was
probably a mistake, as it looks like it is a normal
chromium gtest-based test. This fixes that.

However, it also looks like we're not running this test anywhere,
so I can't easily verify this on a bot. This CL makes no attempt
to fix that, but I did confirm that the test runs locally under
iossim and looks like a gtest-based test.

Bug: 816629
Change-Id: Ibf5370d365e0b7e2ba5238bbcc416db0407d8b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354778
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#798456}
GitOrigin-RevId: a3b72922fa59824e7116123500db35abb8e642f5