Mark chrome_elf_unittests as type 'console_test_launcher'.
At some point initially (I think)? chrome_elf_unittests was
a very thin binary that could not link in //base/test/launcher
to run as a regular chromium gtest-based test.

However, that appears to have changed quite a ways back is a
regular chromium gtest-based test. However, it was still configured
as a "raw" test, which meant that we could only look at the
exit code from the test and not rely on it conforming to the
normal chromium gtest command line API.

This CL fixes that.

Bug: 816629
Change-Id: I184af40d56feb01050a2ddcec020827250a51b7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354664
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#798732}
GitOrigin-RevId: 8895a69e59e5087f7758226225ea326cbb1dc900