[Offline indicator v2] Add command-line switch to force online state
Relands crrev.com/c/2246834 apart from the field trial testing entry and
adds the command-line flag to more tests in test_suites.pyl.

Benchmarks in tools/perf/benchmarks are run on offline devices, and this
causes the offline indicator to show, affecting the stories being run.
Showing the offline indicator isn't realistic when the test is simulating
navigating webpages that would normally require the device to be online.
Also, showing the offline indicator might cause flakiness in the results
or dilute real regressions. We've investigated the reported regressions
and confirmed that they're WAI when the indicator is showing.

This CL adds a command-line switch to prevent the offline indicator from
showing and adds the switch to related benchmarks.

Bug: 1091491
Change-Id: I60a573a527b041ef986a7cddaf5d4705629fc8e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358509
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799366}
GitOrigin-RevId: 9f49c999beb309a1a69c19ecbc5bff0779e337a2