Add --isolated-outdir flag to test executable API.
This adds an optional --isolated=outdir=PATH flag to the
test executable API; this gives the test runners a stable path
to the output directory from a swarming task, so that we can key
other files off of it without needing to pass them all explicitly
via the command line.

This also adds the arg to the isolated_script wrapper, so that we
those types of tests can handle the flag (by ignoring it for now).

We may need to work through the other cases (like gtest tests)
individually, which will be done in subsequent CLs.

Bug: 816629
Change-Id: Ibff6a47544bf94af5d7693a8446dfb669351cc07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358604
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Erik Staab <estaab@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799378}
GitOrigin-RevId: 8744a6783d987c6ba887c1fb9a8ddf87746f977c