Reduce webgl testing load on Mac Pro
Since this bot has low capacity, we can try to keep the testing
load a little lower. We choose to only test the validating command
buffer with webgl2 tests, and the passthrough with webgl1. Also,
we don't need to duplicate SwANGLE testing here since it's likely
to be the same on any Mac.

Bug: chromium:850252
Change-Id: Ib912758efb5e352cc746553f0a3c8d6ee9947cbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363831
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Jonah Ryan-Davis <jonahr@google.com>
Cr-Commit-Position: refs/heads/master@{#803178}
GitOrigin-RevId: 7a19b1d38f4f1c287ae73579553bb77977c07966