Ship Transferable Streams
Remove the feature flag for Transferable Streams and turn it on by
default.

Also remove the web test that I created to stop us shipping it by
accident, since now we are shipping it on purpose.

See the intent to ship thread at
https://groups.google.com/a/chromium.org/g/blink-dev/c/1LStSgBt6AM/m/sJaTciPKBwAJ
and the completed TAG review at
https://github.com/w3ctag/design-reviews/issues/551.

BUG=894838

Change-Id: I56f8891bd1051202ef68bb4623ad8bec53045dff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2428509
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811606}
GitOrigin-RevId: e56b140ca1dad6ffcccaf2e8d8f86b3e99ae71bc