Revert 'infra: Generate source side spec by bucket'.
This CL reverts f60eb053fe03c1896e1abefb5b1da64daf450371
from March of this year; we were generating the big
//testing/buildbot/*.{ci,try}.json files that combined
all of the different buckets/masters, but we weren't
using the files. By reverting the change this reduces
the churn in the repo (and in CLs).

Bug: 1028719
Change-Id: If394897d072836d744f7cd72a2b2b6deb8e06f9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2441651
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#812469}
GitOrigin-RevId: 6269d30fbc65b50fadccc63871d510d8bfe5cbd4