Focus WebContents prior to sending focus to renderer for accessibility
When focus moves from the main WebContents to a guest WebContents, there
currently isn't any mechanism to make sure focus moves to the guest
WebContents. Due to this, the focus call ends up being a no-op which
causes focus not to move along with AT navigation.

Ensure the guest web contents is focused before performing the rest of
the accessibility focus action which allows it to have the desired
effect. This has to happen whether or not the RenderWidgeHostView
believes it is focused (which it does in a guest web contents that
itself does not have focus).

The new test is disabled for ozone wayland - it's passing on all other
platforms (including ozone x11) so will probably need an independent
platform-specific fix (tracked by crbug.com/1137626).

Bug: 1130710

Change-Id: Idd5b9709f255b4085a5268ee52dd6375e88e9412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2446789
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#816808}
GitOrigin-RevId: e1ec154666b2fdeab0cbee3ec31f1d3fbea43770