Change test wrapper script arg handling.
Previously, any isolated script test that ran on a bot
required the `--isolated-script-test-output` flag to be
passed to it.

This CL makes that flag optional.

This was made mandatory originally in order to ensure that
it was being called correctly; however, making it mandatory
means that devs have to set it when using the script locally,
even if they don't care about the output. Fixing the latter
problem seems worth not catching the former at this point;
it seems unlikely we'll introduce a bug in the recipes that
the mandatory flag check would catch.

Also, if you wanted to use a generated bin/run_ script
that contained an argument that would otherwise be expanded
by swarming (something like `--system-log=$ISOLATED_OUTDIR/system.log`),
this CL will allow that to now be embedded into the bin/run_
wrapper. If ISOLATED_OUTDIR is not set locally (and it usually
won't be), then the arg will be dropped.

These two changes in conjunction should allow us to move
all of the command line arguments that are essentially
compile-time-static into the bin/run_ wrapper (thus further
simplifying what is configured in gn_isolate_map.pyl and the
//testing/buildbot/*.pyl files) while still preserving dev-friendliness.

That said, this CL itself should introduce no functional changes
other than the --isolated-script-test-output flag no longer being required.

Bug: 816629
Change-Id: Ic02d2acc44c4a305dceff1fcab50cfe48d515b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2462070
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Luke Z <lpz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821787}
GitOrigin-RevId: 906cfcb89bef663a55b1a9bf8ab3ad8c0d0c0e16