URLPattern: Add BUILD.gn and initial unittest.
This CL adds some initial liburlpattern code and tests.  The main
goal is to stub things out so we can get the build system in place.
The CL does this by creating an initial Parse() API, validates that
input characters are ASCII, and provides a unit test that checks the
case where invalid characters are passed.

Bug: 1141510
Change-Id: I7571d278e2e6d3d2b2a81ffe0e2df677c6831cda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2493246
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821813}
GitOrigin-RevId: ad8343baa0be6dfe819d49e9d0a318cc0c57300b