Enable testKeyboardShownOnDpadCenter on Android P builders
In this test the keyboard was expected to be shown as soon as the
AddressEditor dialog is opened. This behavior was broken because due to
Focus/Keyboard Navigation Changes in Android P+ initial focus is no
longer assigned. The issue was fixed in [1] and now the test should pass
on android-pie-arm64-rel and android-pie-x86-rel builders.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2510211

Bug: 836460, 1061208
Change-Id: I3e857b936b8736b8293f14f2f9fbc82e9bd17b67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2545304
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Elizabeth Popova <lizapopova@google.com>
Cr-Commit-Position: refs/heads/master@{#828778}
GitOrigin-RevId: f6c5e419342131507d5e9e74c4477e539d4dc5fd