Increase browser_tests sharding on Mac ASan
The test suite is timing out on the builder at the moment. This
increases the sharding on the builder, which should allow all tests to
complete, even if it's slow.

The pool is a bit overloaded at the moment, so this won't speed up the
builder, but it will alleviate failures due to timing out.

Bug: 1150555
Change-Id: I3f7d51bbd0206a836da03d036fde4625ed8a0d66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2551454
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Erik Staab <estaab@chromium.org>
Auto-Submit: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/master@{#829391}
GitOrigin-RevId: cd3c383fbb26cd01eea328a5d158fc8aa3dc8190