Referrers: Remove the kReducedReferrerGranularity base::Feature
We've rolled out the new default referrer policy of
strict-origin-when-cross-origin to 100% and are in the process of
cleaning up the corresponding configuration. This change removes the
base::Feature that allowed switching the referrer policy to the new
default, switching the Blink method returning the default
net::ReferrerPolicy object to deterministically return the new default.

It leaves the accessor method (blink::GetDefaultNetReferrerPolicy) in
place, since the method helps let callsites clarify whether they care
about referring to "the default, whatever it is" (in which case they
can call blink::GetDefaultNetReferrerPolicy) or whether they want to
refer to the specific policy that just happens to be the default (in
which case they can use the actual policy value).

Bug: 1150018
Change-Id: I9f246a9cc35877e009f1e79bd822266db620fcb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2545444
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#829805}
GitOrigin-RevId: 776a4b62fb1376be4073509ecaf9998129bd621a