[ios] Enable DragAndDrop by default TOT
Enabling DragAndDrop unexpectedly broke a unit test due to hitting
a DCHECK that had not been hit before.

However, it is a mystery that the unit test was not already broken.
Objects with observers should be shutdown, disconnected, or reset
at the end of the test.

The changes to content suggestions add resets during shutdown so
that the test can pass.

Bug: 1099646
Change-Id: If19a86985eeca0a8bca4810817545dd9bf34091d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2567630
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#832905}
GitOrigin-RevId: d4105ebd73decd003c51d7eeb6bce0254f945a39