ChromeVox Hint in OOBE.
After 20 seconds of inactivity, a hint to activate ChromeVox
will be announced and shown on screen. Design docs can be found in
the crbug.

This change also adds a suite of tests for the feature.

Lastly, this change adds a Chrome OS feature which can be used to
enable/disable the behavior. This is added just in case we want to
toggle the behavior in the future.

ChromeVox will be announced and shown on screen.

Bug: 1124075
Change-Id: I35b82b3c76db89a7f30c2f31029f0d0e29ebd37d
AX-Relnotes: After 20 seconds of inactivity, a hint to activate
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2488081
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/master@{#834270}
GitOrigin-RevId: 03a5179c726ef27d59ee0d41925e07bc821b4119