Enable CriticlaPersistedTabData in fieldtrial_testing_config
This was determined to be necessary in cl/342061697. This is a reland of
chromium CL 2548503. An explanation was provided in the commit
description why loadNextTab() should be called on the UI thread instead
of restoreTabs(false) on the background thread. I deep dived the issue
created by that CL and found that a policy violation occurs when the
directory for CriticalPersistedTabData is created/read. I change the
policy and revert back after the directory is read (and created if
necessary) like we do in several other parts of TabPersistentStore.

Bug: 1151871
Change-Id: I306d4d2bc2ab2b074621d76d1926cc07d1b56b38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2574620
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#834421}
GitOrigin-RevId: 561bac592a9ed3f753304442db080e68bd7450ca