[Extensions] Remove old docserver code
We have a shiny new docserver, and no longer rely on the server in
the Chromium repo. It served us well for many years. Wish it a grand
farewell, and remove it from the codebase.

This removes *everything* under chrome/common/extensions/docs/server2,
as well as the docserver tests introduced in
https://crrev.com/031fdd9a8941a3f6ce45140aad4156f9ddd13606.

Bug: 1161364
Change-Id: Idb18fac9c00decf6c7c46a3f4c1dcb8547d73edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2601680
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#839629}
GitOrigin-RevId: 683a77f65e106bb06d42bab253e6c31efd16aebc