Do not run parallel job for browser_tests on Mac10.13 dbg
InProcessBrowserTest times out on mac. I can't repro this issue locally,
nor the tryjob. When running locally with 4 parallel jobs, I do
occasionally see RunLoop timeout. It's just not hundreds of test failing.
But if I run with 1 parallel job, I don't see any failures.
This cl use 1 parallel job first and see if this can reduce the
flakiness.
This will make the job slower. But as it's not on CQ, it should be
fine to try this out.

Bug: 1152770
Change-Id: I55ff27a506676b1688455c404611a65d643b30fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2601578
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Sven Zheng <svenzheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#841184}
GitOrigin-RevId: b4995e3436e39f285bbe5ebb7ecb7518128cd918