Specify which backend is tested for webgl on ANGLE/Mac CQ
The more specific test suite should be used to ensure we don't miss
testing on any of ANGLE's backends. In this case, the generic test
suite was being used for webgl and webgl2, which defaulted to
passthrough with ANGLE's GL backend.
This CL specifies the validating decoder for webgl1 and passthrough for
webgl2, removing the duplicate testing of webgl1 on the GL backend.

Bug: chromium:982292
Change-Id: I221307922e7fae4f8baa5122abae969ce5d5d4fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2615818
Commit-Queue: Jonah Ryan-Davis <jonahr@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#841654}
GitOrigin-RevId: ce92c8d0714b1a511352045b6c7b02981f27ae6f