Add testing config for per-profile Connectors
A small change in ConnectorsService is required to make this test config
work. When checking analysis/reporting settings, 2 things are required
to obtain settings: a policy set to a valid value and a DM token (from
the browser or the profile). The previous implementation checked the DM
token first, meaning the blocking BrowserDMTokenStorage::RetrieveDMToken
call would trigger on multiple tests and crash them (see patchset 1
failures for examples of this).

The solution is to simply switch the
order of checks and read the policy first. This is also a valid
optimization of Connectors logic, as reading prefs is faster than
potentially reading the DM token from local storage.

Change-Id: Ie1026b57cf7a4fd66663530ec2a77a72f21d6d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2623812
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Reviewed-by: Marc-André Decoste <mad@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843115}
GitOrigin-RevId: 82cf2844165a2ce367c8a78e4d1cbdbe4546c4ea