Add a feature flag for DeclarativeShadowDOM, and enable in fieldtrials
This add a base Feature for DeclarativeShadowDOM, links it to the
Runtime Enabled Feature, and enables this feature in the fieldtrial
testing config.

This feature will likely *not* be launched via Finch, but this will
enable the feature on perf bots. And it will allow a Finch launch,
if desired.

Bug: 1042130
Change-Id: Icd4c2b7c8eab0d85c4552269eb5080c0f70b0189
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2673962
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850316}
GitOrigin-RevId: bd83319957fa53f7d4769fd27b169b74c6bdc2cb