Enable ForceSynchronousHTMLParsing in fieldtrials
NOTE: If a newly-flaky test bisects to this CL, the problem is
likely to be poor assumptions made in the test about when the
test page has been loaded. This CL enables a new synchronous
HTML parsing mode, which can affect the timing of parser yields.
These yield points can vary, per-spec, so if this causes test
instability, likely the test needs to wait, e.g., for something
like window.onload. See the preference/standard/test.js file in
this CL for an example of such a change.

Bug: 901056
Change-Id: Ided734f916e7f98e57b67861750fd0352787257f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705812
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#856868}
GitOrigin-RevId: 66da56bd3f27e9c9643202d12e76ddc6c93323b9