Hide offline link in the network screen when offline policy expired.
When user pods are visible on the login screen and online signin is
enforced for a user there is a way to avoid it by disconnecting networks
and going through the offline flow. This CL blocks this path by checking
this condition on user pod focus.

Bug: b:149562898
Change-Id: Ia78f0dc114460c5de815b458c632ffed14b8edcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2696370
Commit-Queue: Maciek Slusarczyk <mslus@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#857283}
GitOrigin-RevId: 73f243b511b41863ceac650534257eab1319ee1c