Enable (v)python3 in script_test()s.
This CL adds support for running under Python3 (via vpython3)
to the GN script_test() template. This will allow script_test-based
tests to start running under Python3 as you'd expect.

We should be moving script-based tests that currently aren't using
script_test() (and are specifying command lines via gn_isolate_map.pyl)
over to use script_test() as part of the work in crbug.com/816629,
so this change should be viewed as a carrot (or a stick, depending
on how you look at it, I suppose) to that end as well.

This CL changes script_test() to use vpython3 by default (which is
the behavior we'll want going forward), and explicitly updates the
few current script_test targets to still use Python 2.

It also modifies the generate_wrapper() template to contain a
2/3 switch, but leaves the default here to be Python 2, since
there are multiple callers of generate_wrapper.

Bug: 1157665
Change-Id: I4364982f3ab5cd6321dd6d7b411f3620d6d15d4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2715652
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#857531}
GitOrigin-RevId: d7f8c9d6f22171b366c354cf3e93ea9f55158545