Increase sharding on "WebKit Linux layout_ng_disabled Builder"
The trybot is timing out sometimes in builds. This should fix those
issues.

The trybot roughly looks like it has a max of 100 builds a day. Each
task has 1 minute of overhead, so this will generate 8 * 1 * 100 = 800
additional minutes of overhead per day, which should be pretty minor in
the grand scheme of things.

R=gatong, gbeaty

Bug: 1182334
Change-Id: If410211e226522553a10754111ecdc0f0ad76073
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2720661
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#858794}
GitOrigin-RevId: 6eb8a557b9cbdd4a4533a7da80f75e67baf81d6e