Update processing order of metadata builder.
It now does the following:
1. copies checked-in metadata to the output directory
2. [optionally] loads baselines and looks for subtest fails or harness
errors.
3. loads combined statuses from Expectations files in standard blinkpy
order.

Normally, data from #2 and #3 are merged together and clobber whatever
was copied in #1. The exception is Skips which overwrite any other
statuses found in #2 and #3. This is called "translated" metadata.

However, after this change, if an Expectation is annotated with
'wpt_use_checked_in_metadata' in #3 then the statuses from #2 and #3
are discarded and #1 is used. This is called "checked-in" metadata.

This CL also annotates all tests from identity and input bots to use the
'wpt_use_checked_in_metadata' annotation. This uncovered a couple cases
where TestExpectations was covering test failures, and we generated
real metadata for these tests.

Additionally, if we ever use "translated" metadata, we always tell
wptrunner to ignore subtest results, because the translation is unable
to capture this level of granularity.

Bug: 1157848
Change-Id: Ibcabb3a304822d406457515bfe026003e44b62f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2692751
Commit-Queue: Luke Z <lpz@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859023}
GitOrigin-RevId: 50f0fecded229ad6912e5fe65d3dc8c84e980b20