Add test directory -> monorail component metadata to test invocations.
The new ResultDB system has the ability to associate test results with
Monorail components and teams, but in order to do that, we need to
upload the mapping information contained in the testing/DIR_METADATA
files along with the test results.

This CL adds a gclient hook to generate the metadata in a
ResultDB-friendly format that can be accessed by `rdb` and uploaded
as part of a test invocation, and makes the generated data a
dependency of the bin/run_* test scripts, to ensure that every
test invocation will have the data.

Change-Id: I83dcf7f460b6fbd1d4c23616d972510712019148
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2568896
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#861612}
GitOrigin-RevId: 39b25a5f892c66354301b0bd505d45d5a38e6e74