[Lookalikes] Enable target embedding by default on non-iOS platforms
This CL sets the target embedding feature to default enabled in
preparation for the final launch stage.

To avoid false positives, this change also carves out iOS from the list
of supported OSs. The corresponding finch config already didn't apply
to iOS, so this is maintaining the status quo for now. When we fix the
last bit for iOS, we'll roll out target embedding on iOS separately.

Bug: 1076617
Change-Id: I12f8d75299150c1ef89dc74c269997090670275c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2752836
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#862622}
GitOrigin-RevId: 0c547a1d8d09e83fe61ab0ba1be9cc2e1634eb86