Fix 'is not' comparison with literal in testing/xvfb.py
Using 'is' (or 'is not') with a literal is a bug which *usually* works
in CPython (thanks to compiler optimizations) but is not guaranteed to.
Details can be found in [0] but the TL;DR is that 'is' does identity
comparison not value comparison.

[0]: https://adamj.eu/tech/2020/01/21/why-does-python-3-8-syntaxwarning-for-is-literal/

Bug: None
Change-Id: Iaba3f6d88b6725ef90479b1866c4e3e821b7b109
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2756935
Commit-Queue: Dirk Pranke <dpranke@google.com>
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#862888}
GitOrigin-RevId: b4817dc38d963a58caa6aa536c03459c51b810df