[benchmarking] fix the argument config to take the dynamic shard setting
The --use-dynamic-shards argument was set to have type as bool, which will expecte a actual value True/False as the argument. Updated to use it as a switch. When it is not provided, the default is False.

Bug: chromium:1140389
Change-Id: Ifb40125d6d5adb6697296d31cdd139873f8afe24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2762402
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863017}
GitOrigin-RevId: d563d85220f33ee649652f5cc9dc50b48046861a