[ozone/wayland] Guaranteed sub-surfaces to have parent.
Wayland requires sub-surfaces to have parents always.  It should not be
a problem to get a parent when creating a sub-surface, but in a few
places heuristics were used instead of setting parent explicitly, which
caused issues in some situations, including the TooltipControllerTest
suite.

This CL extends constructors of WaylandAuxiliaryWindow and WaylandPopup
with the parent parameter so the sub-surface is guaranteed to have the
correct parent always.  The change fixes tests from
TooltipControllerTest suite.

Bug: 1186974
Change-Id: I0e98e7b85e97dfa6b6f9d88b8e654220053ebaf6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2756530
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Maksim Sisov <msisov@igalia.com>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#866138}
GitOrigin-RevId: fb76a5984eb0f08020375193dc766914c462f777