Document the 'ci_only' flag in //testing/buildbot configs.
The chromium recipe will now support a 'ci_only' boolean flag
on the test suite configs in the //testing/buildbot/ *.pyl
files.

This flag makes it possible for us to indicate that we only want
to run some test suites on the post-submit/CI bots and not on
the pre-submit/trybots. Usually we'll only want to do this when
we have capacity concerns and can't afford to run something in the CQ.

Bug: 1170989
Change-Id: I974cfc7fd31b997c9e0d2731e411fcb389ac26cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2783386
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#866195}
GitOrigin-RevId: b0024a0166edde9659a23a981f181b3e18b232f5