test: add override_board in test template
`override_board` is used in `generate_runner_script` to run
a ChromeOS binary on devices with different target boards,
e.g., run an amd64-generic built binary on an eve device.

This change adds the `override_board` in the test template
so that it can be passed to the `generate_runner_script`.

Bug: 992996
Change-Id: Ief96c8957b637bf47ef5d90454626282e6f6c8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2782028
Commit-Queue: Haoming Chen <hmchen@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#866414}
GitOrigin-RevId: a9e205c1d9ea0594d649258bec9a2deabfd65457