Removed 'Mac11.0 Tests' from chromium.fyi
This is the third of a 3 CL process to rename 'Mac11.0 Tests' to
'Mac11 Tests' and switch the waterfall of the test bot from chromium.fyi
to chromium.mac. In the first step we created a duplicate test bot and
in this step we are removing the original  'Mac11.0 Tests' altogether.

Bug: 1176341
Change-Id: I91ab92e23b7f09dfdae75ed4efad532fa58bd33a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2794816
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Marco Georgaklis <mgeorgaklis@google.com>
Cr-Commit-Position: refs/heads/master@{#867772}
GitOrigin-RevId: 850102b24260d2e8c626296dcf5aeffccce95516