Set app container ACLs for LPAC sandbox.
In order for binaries running under LPAC sandbox to operate they
need read access to their own files.

This change consists of two parts, the first part sets the ACL
when building a component that depends on sandbox (e.g. on a
developer workstation).

The second part sets the ACL when running tests as part of
running tests on a host where the build did not happen e.g.
as part of a swarming isolate or test infrastructure.

Note: on a full Chrome install, these ACLs are set by the
installer.

Add a test to sbox_integration_tests to verify this works.

BUG=1191111

Change-Id: I88d4e53cb11dbd52316cd1a605bce94518f62772
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2811185
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: James Forshaw <forshaw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#874038}
GitOrigin-RevId: 9e4cf31c36e310e9636a5a16d322433f2523c9f8