Delete Java Voice Suggest Provider
This change concludes the experimentation with the Native voice
input and removes the Java voice suggestions provider. The native
provider appears to work as intended and enabled us to get a little
more insight about how Users interact with the Omnibox.

More details available in the summary below:
http://doc/1O-c6YnjEP6x6P--toT1PxeTzEAYdUVnGbqLXLIDJFAY

This change enables us to resume the work on Autocomplete system
reconciliation.

Bug: 1138587
Change-Id: Ic93ce293c194bf88f68da5de531f01a6f38b46ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2831079
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/master@{#874366}
GitOrigin-RevId: ba95bdd732b227324dcb2d8ac7f255cd165a73df