WebLayer: Temporarily remove ToT tests from Android M x86 tester
Need to investigate why the ToT tests are failing on the builder.
We do not need to block adding the skew tests to the main waterfall.
Will re-add the ToT tests to the tester when the problem is fixed.
The ToT tests are still running on android-weblayer-pie-x86-rel-tests so
test results can be viewed there.

Bug: 1183910
Change-Id: Iabc78479288e85e533d307bcb412520070c281fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2841724
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#874752}
GitOrigin-RevId: dae5e8aa47e9d2ff4c67ef5745b7e1188515ecf0