Force NoAllocDirectCall path in WebGL conformance tests w/ fast calls
--enable-fake-no-alloc-direct-call-for-testing means that functions
tagged with [NoAllocDirectCall] will always use the NoAllocDirectCall
implementation, even if Turbofan does not optimize the callsite. This
helps test that the embedder does not violate any of the
NoAllocDirectCall restrictions inside the callback, without relying on
Turbofan to optimize a callsite for the code path to be taken.

Bug: 1187263
Change-Id: I81b3ab5f8c5082bcfc5cee35e60e996bdedbbf28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2872245
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#879145}
NOKEYCHECK=True
GitOrigin-RevId: 6b5926040afe51ed90bfffbdb2fed3b0a225dae2