Move more tests to faster Mac11 machines.
I'm still attempting to tune the sharding and timeouts for our
Mac11 capacity, and it looks like unit_tests and content_unittests
are too slow to run on the "slow" Mac11 machines (the so-called
'n1-highcpu-4' machine type, which is actually one-third the
speed of the 'n1-standard-4' machine type).

So, this CL moves them over to run on the faster machines.

Bug: 1206401
Change-Id: I5b33823892d14f5100e3d2c06f7f5170d4004992
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2878107
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#880512}
NOKEYCHECK=True
GitOrigin-RevId: 4101a281695d7d628f75a6431360c0a179e4d9a8