[ozone/wayland] Fixed tests from the RemoteCopyBrowserTest suite.
The RemoteCopyBrowserTest test suite in browser_tests was disabled on
Wayland bots.  It turned out that the Wayland implementation had an
unnecessary CHECK that was hit.  The X11 implementation of the same
method doesn't have that assertion.

This patch removes the aforementioned CHECK and enables tests on
Ozone/Wayland.

Bug: 1176174
Change-Id: I28e3d73c6186cbb77f253db35c6bc8ea85c9a836
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2894244
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#882867}
NOKEYCHECK=True
GitOrigin-RevId: 1d7165b5650368dd0b0ca4c7e9af0f312df54f02