[ozone/wayland] Enabled tests that needed XKB support.
Initially it was assumed that XKB is not set up in the configuration
used on the test bots.  However, it turned out that the real issue is
related to threading.  The keymap is supposed to be set up at the stage
of initialisation, and when the standalone Weston is used,
XkbKeyboardLayoutEngine::SetKeymap() is called early, before entering
the test body.  However, When the bundled Weston is used, that call is
delayed and happens *after* the test body (and the test fails because it
tries to use XKB before it has got the state).

This patch fixes the issue by adding a closure that the keyboard layout
engine calls back when it is ready to accept calls that require the
keymap.  Tests that require XKB may now wait until it is ready.

Bug: 1186996
Change-Id: I53c543ce7fb7dbfcd4a34b613b8df63f947bd6e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2878964
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Kevin Schoedel <kpschoedel@chromium.org>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#884902}
NOKEYCHECK=True
GitOrigin-RevId: 5e34d8e72c191720f964067b283aed0d85e0f442