Move metrics_python_tests from isolated_script to script.
This changes how we invoke metrics_python_tests, so that
it is invoked as a 'script' test and run directly on the builder
(machine doing the compile) rather than as an 'isolated_script'
test under swarming.

This fixes an issue where we'd try to run `dirmd`, which requires
a checkout.

Bug: 1201822
Change-Id: I17c4f6c2341c70b688efc48aada88a5a51ef4a16
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2921324
Auto-Submit: Dirk Pranke <dpranke@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893147}
NOKEYCHECK=True
GitOrigin-RevId: 4c9955b1c31b725372e61590c595522cb4e63451