Fix flakes for dialog pixel tests
A lot of flakes are caused by mouse interfering with the dialog and
hovering over a button or inconsistencies in focus behaviour. Usually,
dialogs have a default focused view but sometimes when running
tests in parallel, dialog is inactive and the default view isn't
focused.

This CL fixes both of those problems by 1. moving the mouse in 0x0
position and 2. clearing focus of the view before taking screenshots.
This should remove all the incosistencies and fix the flakes.

Known not fixed flakes:
1. IPH tests flake, sometimes screenshots include the anchor arrow,
sometimes not.
2. Chip test flakes on the size of screenshot taken. Sometimes it cuts
out the text and include only the icon part of the chip. First theory
was that it is caused by animation triggering and chip starting to
collapse but adding a setter to force chip to stay expanded didn't help.

Bug: 1208502
Change-Id: I626f988ae448a3a35ae57e055dcc98bcbf7dd33d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3034889
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Cr-Commit-Position: refs/heads/master@{#904917}
NOKEYCHECK=True
GitOrigin-RevId: 996693cdebaa2e07cc853570a7f46ec659a7f8a8