Add variant to unexpected pass finder data types
Adds the concept of a variant to the unexpected pass finder's Result
and Expectation classes. This is currently not needed, but will be for
the web tests, as there are multiple flag-specific variants of the tests
which need to be differentiated.

Bug: 1222827
Change-Id: Ibe62359add8a98df4be0ef5b22c2c9193e55b00d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3063690
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#907292}
NOKEYCHECK=True
GitOrigin-RevId: 146ffb9b8b51b747740b85367576abaac187c9b0