Add system-log-template to performance tests for per-benchmark sys logs.
When multiple suites are run on the same shard, the system logs for that
run are overwritten/specified in the same location (for Cr-Fuchsia
telemetry runs). This would allow system logs to be specified in the
benchmark directory for each run, so they're available instead of being
overwritten.

Bug: 1236245
Change-Id: I2f641d936b138c7e301d88b033d4ba45554159a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3083302
Commit-Queue: Rohan Pavone <rohpavone@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910505}
NOKEYCHECK=True
GitOrigin-RevId: aa882fd7826a8a8f4f770e927c4bde3201847e08