Disable UseVideoFrameSinkBundle for Android tests
There's a bug affecting Android which causes lots of I/O churn and
therefore large CPU regressions on some perf benchmarks. This experiment
isn't running on Android yet anyway, so it's a mistake that the config
was turned on for that platform already. This turns it off.

Bug: 1240464
Change-Id: If58ab9634c0ec12d8c0f6350fd05cbd5f3b6aca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3105346
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#913530}
NOKEYCHECK=True
GitOrigin-RevId: 4c4d157c17d7d3cfef6038d4e321d889b4a01235