lacros: Use AccountManager in browser tests
Use Ash AccountManager in Lacros tests, when crosapi is disabled.
Re-enable some of the browser tests.

Browser tests which depend on Account Manager are failing on Lacros.
They try to update credentials in Ash Account Manager, which is currently
not possible because of lack of crosapi. To run these tests on Lacros
are creating an Ash AccountManager instance in tests and talking to it
in-memory.

- Change account_manager_facade_factory_lacros to return an instance of
`AccountManagerFacade` that is connected to `AccountManagerMojoService`
for tests, if `CreateAshAccountManagerForTests()` was called before.
- Create an instance of Ash AccountManager in IdentityBrowserTestBase.
- Update identity_test_utils to use AccountManager on Lacros if account
consistency is `kMirror` (profiles with DICE account consistency use
`ProfileOAuth2TokenService` for managing accounts).

Bug: 1197201
Change-Id: Icae7475adefc7f849854a2e078c225fa0cc878cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3000145
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919694}
NOKEYCHECK=True
GitOrigin-RevId: 4382044a1ba778fb5adca48013f3d6a657cd7252