Add the new UserAgentStringOrdering test name to the Fuchsia filter file
In crrev.com/c/3167270, UserAgentUtilsTest was changed to become a
parameterized test, but we didn't update the Fuchsia expectations file.
This caused the UserAgentStringOrdering test to fail on Fuchsia CI.

This CL updates the expectations test file to account for the new test
name.

Bug: 1251660
Change-Id: Ia1c080ff40b0ee102a7bc4d6405255a6e91bf78c
Fixed: 1251660
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3173560
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Ali Beyad <abeyad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923866}
NOKEYCHECK=True
GitOrigin-RevId: 75a98a090d300969118150c23620fbe0251e2963