Remove the non-LPAC sandbox code on Windows.
The restricted token sandbox will not ship, as the LPAC sandbox
has been chosen as the sandboxing solution on Windows.

This CL removes the restricted token sandbox code, and disables
the sandbox on fieldtrial_testing_config because it's not yet
ready to ship.

This CL also updates the supported version of Windows for the
LPAC sandbox to WIN10_RS1 (Version 1607, Build 14393) to match
the support offered by the sandbox.

BUG=841001

Change-Id: I9b32caf141b32e75992f8b802138b11dd10deb3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3171280
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#924583}
NOKEYCHECK=True
GitOrigin-RevId: 46eddccdd3af1caf4b8fd785afb0d69a610f3748