Retire the OmniboxCompactSuggestions flag.
The purpose of this flag is to only slightly reduce the height of
the suggestion view padding. The semi-compact variant is being
released as part of the Adaptive Suggestions feature.

This flag is fully launched as part of Extended ZPS and enables the
second of the two features involved (the other feature being enabled
here: https://crrev.com/c/3169417)

Bug: 1123200
Change-Id: Icbdc7dafb7a1bf2ad3fc2adb2350e13d1da9969b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3194377
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#926982}
NOKEYCHECK=True
GitOrigin-RevId: 1bdf658163d427935d890f4157c13c674174b661