Add ForceSynchronousHTMLParsing and LoaderDataPipeTuning to fieldtrials
====== Sheriffs! ======
If you believe this CL is causing widespread flakiness, feel free to
revert it. However, if you believe a handful of tests were made flaky
by this CL, I'd much prefer if you disabled those tests and filed bugs
against the test owners, rather than reverting this entire CL, which
took some work to land. You can include any such tests within the
TestExpectations section marked like this:
   # === Synchronous, budgeted HTML parser tests from here ====
Thanks!
=======================

We'd like to start a Finch trial of the ForceSynchronousHTMLParsing
feature, along with the associated buffer/chunk size changes, in
Stable. As such, this CL adds the test configs to fieldtrial testing
config. The corresponding GCL is located here:

http://google3/googledata/googleclient/chrome/finch/gcl_studies/LoaderDataPipeTuning.gcl

This CL changes the default timing of parser yields, which affects
the timing (but *not* the correctness of timing) during tests. Some
tests are incorrectly (and likely inadvertently) sensitive to the
timing of yield/render points, and fail with this flag enabled.
However, these are all bugs in the test, rather than a problem with
this feature. As a result, I've disabled the affected tests, and filed
bugs for each, to manage removing the dependence and re-enabling the
test. If you get here from a test that you believe to be working
correctly, and you suspect that this CL is triggering incorrect
behavior, please contact me ASAP!

Additionally, several tests change their output due to the yield
timing change, but the new output is also perfectly valid. Those tests
are rebaselined here. I have reviewed all baselines in this CL and
all are trivial output changes, such as changes to CONSOLE messages,
or changes in unrelated output line order.

And for a few simple tests, I just fixed the test flakiness myself:
  fast/css/focus-display-block-inline.html
  .../remove-from-grouped-mapping-on-reassignment-crash.html
  paint/invalidation/compositing/inline-repaint-container.html

A particular note about the line numbers that sometimes appear in
CONSOLE messages: those are already flaky. See crbug.com/896194.
I don't believe this CL made that situation any worse, but it did
appear to change the timing a bit. So any newly-discovered flaky
tests due to console line numbers should be assigned to
crbug.com/896194.

Bug: 1041006,896194,1254921
Change-Id: I8398cc522cf4a76118f0a6bdc93f9a0ce4743514
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3187376
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Richard Townsend <richard.townsend@arm.com>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927910}
NOKEYCHECK=True
GitOrigin-RevId: 46f4a57c31bc0d5cda7fb93be3f4f92d99a76dbc