ozone/wayland: add try config to wayland bot and make bot run tests
This is a follow up for https://crrev.com/c/3194070 and
https://crrev.com/c/3188520 that adds a (missing)
(and yet optional) try config for linux-wayland-rel
bot and also makes CI and optional CQ wayland bot run
chromium_linux_gtests. Failures are not expected as the config
is the same as linux-ozone-rel has with the only difference that
the bot becomes a Wayland only one.

Bug: 1252668
Change-Id: I04eb6dacff3ca8a91bbeacf054e0f6afab071b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198173
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928025}
NOKEYCHECK=True
GitOrigin-RevId: 4681124cd1ffbbe0ebac6834d5d49d1a02fc8fc5