OOBE: Add pre-enrollment TPM check for devices with dynamic TPM
It is possible that TPM isn't owned by the OS installed on the device.
This CL introduces a pre-enrollment check that makes sure that the TPM
is set up correctly during enrollment. It only applies to the devices
with --tpm-is-dynamic switch enabled.

Bug: b:190966258
Change-Id: Ibee174745ed9cd9b32d0b80cb3ec174328e94e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3172616
Commit-Queue: Danila Kuzmin <dkuzmin@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928101}
NOKEYCHECK=True
GitOrigin-RevId: 8f101daf8a7a43656033bf091cc109aae542d7f6