base/allocator: Allow one more static initializer for macOS
When PartitionAlloc-Everywhere gets enabled for macOS, PA-E uses
a static initializer (InitializeDefaultMallocZoneWithPartitionAlloc),
so increase the limit of number of static initializers by one.

We're aware that the static initializer shouldn't be the best option
to install our own malloc zone.  However, in order to unblock
MiraclePtr project, we decided to go this way as a temporary solution.

Pros/cons of the static initializer, other possible options, etc.
were discussed offline, and the discussion is summarized at:
https://docs.google.com/document/d/1qpEIeqk3whkwOXcKfpwVafJu859vFmMh75XJdbCo6vs/edit?usp=sharing

We'll improve the interception mechanism in the future.  Especially,
ParitionAlloc library (as a standalone library) is promising and
on-going.

Change-Id: I73a8fdd36dcac65466419824437e59aa7d2579d5
Bug: 1121427
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3143794
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#932521}
NOKEYCHECK=True
GitOrigin-RevId: bba574cfb0e924321c2391b96f08d759faf48e2a