Fix All/CertVerifyProcInternalWithNetFetchingTest.Sha1IntermediateButAIAHasSha256/CertVerifyProcAndroid failures.
Newer Android versions reject the chain due to the SHA1 intermediate,
but do not build the correct chain by AIA. Since only the partial
chain is returned, CertVerifyProc does not mark it as SHA1 as it does
not examine the last cert in the chain.

Update the test to such that on Android if ERR_CERT_AUTHORITY_INVALID is
returned, don't check the rest of the statuses.

Fixed: 1191795
Change-Id: I21d06ac08a4109ec94fc845ed090175a2956d43e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3255965
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#937527}
NOKEYCHECK=True
GitOrigin-RevId: c836ef86f53f0b46262daef26ed93f80c8c80798