Enable viz_unittests, cc_unittests on android-pie-arm64-rel CQ
Running them on x86 does not provide enough coverage, as they are
important to run on hardware that is close to real Android device.

The test runtimes are 3 mins for viz and 7 mins for cc which are fairly
fast so the current P device tests pool (http://shortn/_U0a5Ul0S9O)
should be able to handle it.

Bug: 1266134
Change-Id: If83bc30f281d9b3c388172afd276bcf70f94e90d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3260765
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#938845}
NOKEYCHECK=True
GitOrigin-RevId: 5e9fa6cbb125776c7d8d24f2cfb9b21b69855073