Run check_network_annotations.py on Android CQ
Since the Network Traffic Annotation Auditor isn't 100% ready yet,
for now this means running the Auditor in "FYI mode". i.e., it
prints error to stdout, but it's green on CQ. That way, it doesn't
block submitting CLs if it's buggy.

This will let us debug issues, to eventually turn it on for real.

Bug: 1254719
Change-Id: I52a2b6a4ba9803bb85b07403b38d11791faadbb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3258203
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#939987}
NOKEYCHECK=True
GitOrigin-RevId: 8b337c23057c824b98afeeb21bb2c7e7b655ff8b