Allow chrome binary args in variations tests
I need to pass an argument to the chrome binary to tell it which channel
to run as. I assume that other flags might be needed in the future, so
just passing arguments through could be useful. I could be wrong though.

Bug: 1269554
Change-Id: Ia622ed5f9b081e3ca17514106a13e0f3a1ecb8e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3277088
Reviewed-by: Arthur Wang <wuwang@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942777}
NOKEYCHECK=True
GitOrigin-RevId: cfb597d87fee0ce471c28a693e0b4d24454ee691