[Autofill Assistant] Enable animations + Remove filter
In crrev/c/3309117 we disabled the tests for the "emulator_p".
This change reverts this and force-enables animations.

Note: It is unclear why this is required - it should not be.
Tests are really flaky (or totally broken) without.

Warning: This has the potential to make the builder(s) flaky
again. Revert in case of doubt.

Bug: 1272997
Bug: b/208965862
Change-Id: I5cd3f01592c80f0ed942face42d4904c6f769f36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3338359
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Luca Hunkeler <hluca@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/main@{#951857}
NOKEYCHECK=True
GitOrigin-RevId: fd041bd289c023c7b07befdc3b9315620e833037