[LR] Add fieldtrial config & fix tests
Adds a fieldtrial testing config entry for Lightweight Reactions in
preparation for ramping up the experiment to Beta.

This also updates a bunch of sharing hub tests to account for the new
Lightweight Reactions entry point.

This is a reland of crrev.com/c/3331121, which had been reverted due
to the sharing hub tests failing when the LR feature was enabled by the
trial config.

Change-Id: I685aeec51fda2e7eefe79949b057617757e4b079
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3345133
Reviewed-by: Sophey Dong <sophey@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/main@{#952968}
NOKEYCHECK=True
GitOrigin-RevId: dd667778bddb374023a97161c465988fdf522662