Remove ClankStartupOptimizations from fieldtrial_testing_config
We have turned down this experiment as certain arms of it interplay
negatively with the change of Java BOOTSTRAP tasks back to having
USER_VISIBLE priority. This CL similarly removes it from
fieldtrial_testing_config.json to (a) avoid that negative interplay
impacting tests, and (b) put the perfbots in a consistent state wrt
production behavior.

This CL is a reland; the original change was unintentionally undone in
a rebase of an unrelated CL [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3335077

Bug: 1280192
Change-Id: I05e3863424fc37d2e7edff3918faec37a566dfdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3365213
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#955113}
NOKEYCHECK=True
GitOrigin-RevId: 9962bdc9633dd1960bc98c3b1cc44211ca0f04e9