Use fake network connection tracker in browser tests on Fuchsia
The real network connection tracker never indicates that the browser is
online for Fuchsia. These tests don't actually care about online/offline
(that behavior is tested in the AttributionManagerImpl unit tests), so
we use a fake tracker that always indicates that the browser is online.

If the behavior on Fuchsia is ever changed, we will go back to using the
real tracker.

We change the AttributionManagerImpl to avoid injecting the tracker, and
instead retrieve it via content::GetNetworkConnectionTracker(), which
allows us to override it in attributions_browsertest.cc more easily via
content::SetNetworkConnectionTrackerForTesting().

Prior to crrev.com/c/3318796 these tests didn't fail because the
reporter assumed that the browser was online unless it heard otherwise,
whereas now the manager always queries the actual online status.

Bug: 1285057
Change-Id: Ifa765eab8435a5a4dc92631a8639096bc874b8a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3373141
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Chong Gu <chonggu@google.com>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957390}
NOKEYCHECK=True
GitOrigin-RevId: dcf0d8b867a30928ddfcb4f027e131ca3961f255