What's New: Remove CrOS from the field trial testing config
Launch on CrOS is being delayed for now, and this seems to be causing
flakes on some M98 CrOS bots. Rather than reverting a release blocker
fix as is currently proposed, remove CrOS from the field trial config
so that such bots won't run with the feature enabled.

This is safe since we are not currently launching the feature on CrOS.

Bug: 1278962
Change-Id: Ie108f6e6cfad23e5d2b59dfa01af8f6b73ba1b0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3384720
Reviewed-by: Mohamad Ahmadi <mahmadi@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958719}
NOKEYCHECK=True
GitOrigin-RevId: ea7749668bebcfd272b0fe39b54d5f432bc22344